Conditionally-imported extra tests for Python 3
The Python 3-specific tests for keyword-only argument handling will cause
compilation of the test suite to fail on Python 2, since Python 2 doesn't
understand keyword-only argument syntax.  So we create a separate module
for Python 3 tests which is conditionally imported into the main test
suite, based on the current Python version.