Don't initialize for classmethod/staticmethod objs
Unbound class and static methods, in the form of classmethod or
staticmethod instances, are not callable, so don't bother initializing
them.  Raise an appropriate error if someone attempts to call one of these
partial instances for some reason (rather than allow __call__ to continue
on and fail because the object wasn't initialized, producing a more obscure
error).